+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `CSSInliner` and customization options. [#9](https://github.com/Stranger6667/css-inline/issues/9)
+- Option to remove "style" tags. [#11](https://github.com/Stranger6667/css-inline/issues/11)
 
 ### Changed